Extract BRSessionClient into separate file
