+++ BeeRacingTests/BeeRacingConnectionTests.swift
@@ -8,22 +8,6 @@
 import XCTest
 @testable import BeeRacing
 
-class BRSessionClient {
-    private let session: BRSessionProtocol
-    
-    public init(session: BRSessionProtocol = URLSession.shared) {
-        self.session = session
-    }
-    
-    public func performRequest(from url: URL) async throws -> Data {
-        guard let (data, _) = try? await session.get(from: url, delegate: nil) else {
-            throw BRSessionError.connectivity
-        }
-        
-        return data
-    }
-}
-
 final class BeeRacingConnectionTests: XCTestCase {
 
     func testDoesNotPerformAnyURLRequest() {
